Isolate unit test covering custom methods of storage Entity Account
